related to lesson 5

+++ database/migrations/2020_01_05_114709_create_project_table.php
@@ -18,7 +18,7 @@ class CreateProjectTable extends Migration
             $table->timestamps();
             $table->string('title');
             $table->text('description');
-            $table->unsignedInteger('owner_id');
+            $table->unsignedBigInteger('owner_id');
             $table->foreign('owner_id')->references('id')->on('users')->onDelete('cascade');
         });
     }
